Move entityAttributeText
